Code Architecture Brainstorming ended like this
This is the architecture the team had in mind before starting to code. This is a very rough image of the architecture but this is how it ended as well. The presentation layer was at first for a console application, which could then be replaced by a web presentation layer without having to redesign everything (the other layers still useful).